# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skim",
    "version": "1.0.0",
    "description": "Skim \u2014 clean web reader for AI agents. Turns any URL into agent-ready Markdown (no nav, no ads, no boilerplate) plus structured metadata, in milliseconds \u2014 ~4x smaller than raw HTML, so agents pay for ~4x fewer input tokens. Default path: free API key (sk402_, 1,000 credits, no wallet) on /api/t/* endpoints. Optional path: pay per call in USDC via x402.",
    "x-guidance": "Recommended: create a free sk402_ API key at https://skim402.com and call GET /api/t/read with Authorization: Bearer sk402_\u2026. Token-gated endpoints live under /api/t/*. Optional x402 path: Skim also exposes GET /api/v2/read (single URL \u2192 markdown), GET /api/v2/read/js (JS-rendered pages), POST /api/v2/read/batch (up to 10 URLs in one paid call), POST /api/v2/extract (URL + JSON Schema \u2192 schema-conforming JSON via LLM). Those endpoints return 402 Payment Required with x402 v2 payment details; sign the EIP-3009 USDC authorization with any x402 v2 client, retry, and you get the result. Pricing: $0.002 basic read, $0.005 JS render, $0.02 batch (10\u00d7 single), $0.015 structured extract \u2014 all settled in USDC on Base mainnet on the wallet path. Settlement happens only on 2xx responses, so failed reads are never charged. Identical v1 POST endpoints remain available for legacy clients at /api/v1/*.",
    "contact": {
      "name": "Skim",
      "url": "https://skim402.com"
    }
  },
  "servers": [
...
  "paths": {
...
    "/api/t/read": {
      "get": {
        "operationId": "tokenReadUrl",
        "summary": "Read a URL (card-lane token auth)",
        "description": "Same clean reader as /api/v2/read, paid with plan/pack credits (1 credit; automatic browser fallback costs 1 extra credit). Failed reads are refunded.",
        "tags": [
          "card"
        ],
        "security": [
          {
            "skimToken": []
          }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "The fully-qualified http(s) URL of the page to read."
          }
        ],
        "responses": {
          "200": {
            "description": "Clean reader output",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown token"
          },
          "402": {
            "description": "Out of credits"
          },
          "422": {
            "description": "URL could not be fetched or had no readable content"
          }
        }
      }
    },
    "/api/t/read/js": {
      "get": {
        "operationId": "tokenReadUrlJs",
        "summary": "Browser-rendered read (card-lane token auth)",
        "description": "Forced headless-browser render, same response shape as /api/t/read. Costs 3 credits; failed reads are refunded.",
        "tags": [
          "card"
        ],
        "security": [
          {
            "skimToken": []
          }
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "The fully-qualified http(s) URL of the page to read."
          }
        ],
        "responses": {
          "200": {
            "description": "Clean reader output",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadResult"
                }
              }
            }
          },
          "401": {
            "description": "Missing or unknown token"
          },
          "402": {
            "description": "Out of credits"
          },
          "422": {
            "description": "URL could not be fetched or had no readable content"
          }
        }
      }
    },
...
    "/api/t/extract": {
      "post": {
        "operationId": "tokenExtract",
        "summary": "Schema-constrained extraction (card-lane token auth)",
        "description": "Same extraction as /api/v2/extract. Costs 8 credits when at least one usable row or object is returned. Empty extracts (tables:[], rows:[], items:[], \u2026) return 422 and are not charged; if credits were already deducted they are refunded. Other failed extractions are refunded.",
        "tags": [
          "card"
        ],
        "security": [
          {
            "skimToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "url",
                  "schema"
                ],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "schema": {
                    "type": "object",
                    "description": "JSON Schema the output must match"
                  },
                  "instructions": {
                    "type": "string",
                    "maxLength": 2000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Typed JSON matching the supplied schema. charged=8 when usable rows/objects are present."
          },
          "401": {
            "description": "Missing or unknown token"
          },
          "402": {
            "description": "Out of credits"
          },
          "422": {
            "description": "URL could not be fetched, had no readable content, or extraction returned no usable rows (not charged)"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ReadResult": {
        "type": "object",
        "required": [
          "url",
          "finalUrl",
          "mode",
          "markdown",
          "text",
          "metadata"
        ],
        "properties": {
          "url": {
            "type": "string",
            "description": "The URL as supplied by the caller."
          },
          "finalUrl": {
            "type": "string",
            "description": "URL after redirects."
          },
          "mode": {
            "type": "string",
            "enum": [
              "basic",
              "js"
            ],
            "description": "Reader mode used to produce this result."
          },
          "markdown": {
            "type": "string",
            "description": "Cleaned article body as Markdown."
          },
          "text": {
            "type": "string",
            "description": "Plain-text fallback."
          },
          "metadata": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "nullable": true
              },
              "byline": {
                "type": "string",
                "nullable": true
              },
              "siteName": {
                "type": "string",
                "nullable": true
              },
              "publishedAt": {
                "type": "string",
                "nullable": true,
                "description": "Published date as supplied by the page (typically ISO 8601)."
              },
              "excerpt": {
                "type": "string",
                "nullable": true
              },
              "lang": {
                "type": "string",
                "nullable": true
              },
              "length": {
                "type": "integer",
                "nullable": true,
                "description": "Approximate character length of the extracted article body."
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "skimToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "sk402_...",
        "description": "Card-lane API token issued at checkout (skim402.com/card). Shown once on the success page. Also accepted via the x-skim-token header."
      }
    }
  }
}